# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/contactos": {
      "post": {
        "operationId": "ContactosController_crear",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateContactoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Crea un contacto (cliente, proveedor, o ambos) dentro de la organización activa.",
        "tags": [
          "Contactos"
        ]
      },
      "get": {
        "operationId": "ContactosController_listar",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Búsqueda ILIKE parcial sobre razonSocial y nombreComercial (OR). Usa GIN trigram.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "documento",
            "required": false,
            "in": "query",
            "description": "Match exacto sobre documento",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "esCliente",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "esProveedor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "activo",
            "required": false,
            "in": "query",
            "description": "true (default) → solo activos; false → solo inactivos; all → activos + inactivos.",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 100,
              "default": 50,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListarContactosResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Lista contactos del tenant con paginación y filtros (q, documento, esCliente, esProveedor, activo).",
        "tags": [
          "Contactos"
        ]
      }
    },
    "/api/contactos/{id}": {
      "get": {
        "operationId": "ContactosController_obtener",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Detalle del contacto.",
        "tags": [
          "Contactos"
        ]
      },
      "patch": {
        "operationId": "ContactosController_actualizar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateContactoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "PATCH del contacto. Campos opcionales; sólo toca los que vienen. El toggle activo/inactivo vive en endpoints dedicados.",
        "tags": [
          "Contactos"
        ]
      },
      "delete": {
        "operationId": "ContactosController_eliminar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Elimina un contacto. Falla con 409 si alguna línea de comprobante lo referencia (usar desactivar en ese caso).",
        "tags": [
          "Contactos"
        ]
      }
    },
    "/api/contactos/{id}/desactivar": {
      "post": {
        "operationId": "ContactosController_desactivar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Marca el contacto como inactivo. Idempotente. Los comprobantes históricos siguen refiriéndolo; los nuevos no pueden usarlo.",
        "tags": [
          "Contactos"
        ]
      }
    },
    "/api/contactos/{id}/reactivar": {
      "post": {
        "operationId": "ContactosController_reactivar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Reactiva un contacto inactivo. Idempotente.",
        "tags": [
          "Contactos"
        ]
      }
    },
...
    "/api/libros/diario": {
      "get": {
        "operationId": "ReportesController_obtenerLibroDiario",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibroDiarioResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Libro Diario: listado cronológico de asientos CONTABILIZADOS y BLOQUEADOS. Filtrar por periodoFiscalId O fechaDesde+fechaHasta. Tope: 5.000 asientos (REQ-LD-10).",
        "tags": [
          "Libros contables"
        ]
      }
    },
    "/api/libros/mayor": {
      "get": {
        "operationId": "ReportesController_obtenerLibroMayor",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibroMayorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Libro Mayor: vista por cuenta con saldo inicial, running balance y saldo final. Filtrar por periodoFiscalId O fechaDesde+fechaHasta. Filtro opcional por cuentaId. Tope: 20.000 movimientos (REQ-LM-12).",
        "tags": [
          "Libros contables"
        ]
      }
    },
...
  },
  "info": {
    "title": "Multi-Tenant SaaS API",
    "description": "Production-ready NestJS API with multi-tenancy, authentication, RBAC, billing, and more.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter JWT token"
      },
      "X-Tenant-ID": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Tenant-ID",
        "description": "Tenant ID for multi-tenant operations"
      }
    },
    "schemas": {
...
      "CreateContactoDto": {
        "type": "object",
        "properties": {
          "razonSocial": {
            "type": "string",
            "example": "Granjas El Sol SRL",
            "minLength": 2,
            "maxLength": 200
          },
          "nombreComercial": {
            "type": "string",
            "example": "El Sol",
            "maxLength": 200,
            "nullable": true
          },
          "documento": {
            "type": "string",
            "example": "1234567019",
            "maxLength": 50,
            "nullable": true,
            "description": "NIT, CI, CEX o pasaporte. Texto libre."
          },
          "esCliente": {
            "type": "boolean",
            "example": true
          },
          "esProveedor": {
            "type": "boolean",
            "example": false
          },
          "email": {
            "type": "string",
            "example": "ventas@elsol.bo",
            "nullable": true
          },
          "telefono": {
            "type": "string",
            "example": "+591 3 1234567",
            "maxLength": 50,
            "nullable": true
          },
          "direccion": {
            "type": "string",
            "maxLength": 500,
            "nullable": true
          }
        },
        "required": [
          "razonSocial",
          "esCliente",
          "esProveedor"
        ]
      },
      "Object": {
        "type": "object",
        "properties": {}
      },
      "ContactoResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "razonSocial": {
            "type": "string"
          },
          "nombreComercial": {
            "type": "string",
            "nullable": true
          },
          "documento": {
            "type": "string",
            "nullable": true
          },
          "esCliente": {
            "type": "boolean"
          },
          "esProveedor": {
            "type": "boolean"
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "telefono": {
            "type": "string",
            "nullable": true
          },
          "direccion": {
            "type": "string",
            "nullable": true
          },
          "activo": {
            "type": "boolean"
          },
          "createdByUserId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "razonSocial",
          "nombreComercial",
          "documento",
          "esCliente",
          "esProveedor",
          "email",
          "telefono",
          "direccion",
          "activo",
          "createdByUserId",
          "createdAt",
          "updatedAt"
        ]
      },
      "ListarContactosResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContactoResponseDto"
            }
          },
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "pageSize": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "pageSize"
        ]
      },
      "UpdateContactoDto": {
        "type": "object",
        "properties": {
          "razonSocial": {
            "type": "string",
            "minLength": 2,
            "maxLength": 200
          },
          "nombreComercial": {
            "type": "string",
            "maxLength": 200,
            "nullable": true
          },
          "documento": {
            "type": "string",
            "maxLength": 50,
            "nullable": true
          },
          "esCliente": {
            "type": "boolean"
          },
          "esProveedor": {
            "type": "boolean"
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "telefono": {
            "type": "string",
            "maxLength": 50,
            "nullable": true
          },
          "direccion": {
            "type": "string",
            "maxLength": 500,
            "nullable": true
          }
        }
      },
...
      "RangoFechasDto": {
        "type": "object",
        "properties": {
          "fechaDesde": {
            "type": "string",
            "example": "2026-04-01"
          },
          "fechaHasta": {
            "type": "string",
            "example": "2026-04-30"
          }
        },
        "required": [
          "fechaDesde",
          "fechaHasta"
        ]
      },
      "LineaLibroDiarioDto": {
        "type": "object",
        "properties": {
          "codigoCuenta": {
            "type": "string"
          },
          "nombreCuenta": {
            "type": "string"
          },
          "glosa": {
            "type": "string",
            "nullable": true
          },
          "debeBob": {
            "type": "string",
            "example": "1000.00"
          },
          "haberBob": {
            "type": "string",
            "example": "0.00"
          }
        },
        "required": [
          "codigoCuenta",
          "nombreCuenta",
          "glosa",
          "debeBob",
          "haberBob"
        ]
      },
      "AsientoLibroDiarioDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "fechaContable": {
            "type": "string",
            "example": "2026-04-22"
          },
          "numero": {
            "type": "string",
            "nullable": true,
            "example": "I2604-000042"
          },
          "tipo": {
            "type": "string"
          },
          "estado": {
            "type": "string"
          },
          "glosa": {
            "type": "string"
          },
          "anulado": {
            "type": "boolean"
          },
          "lineas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LineaLibroDiarioDto"
            }
          }
        },
        "required": [
          "id",
          "fechaContable",
          "numero",
          "tipo",
          "estado",
          "glosa",
          "anulado",
          "lineas"
        ]
      },
      "LibroDiarioResponseDto": {
        "type": "object",
        "properties": {
          "rango": {
            "$ref": "#/components/schemas/RangoFechasDto"
          },
          "asientos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AsientoLibroDiarioDto"
            }
          },
          "totalDebeBob": {
            "type": "string",
            "example": "1000.00"
          },
          "totalHaberBob": {
            "type": "string",
            "example": "1000.00"
          }
        },
        "required": [
          "rango",
          "asientos",
          "totalDebeBob",
          "totalHaberBob"
        ]
      },
      "MovimientoMayorDto": {
        "type": "object",
        "properties": {
          "comprobanteId": {
            "type": "string"
          },
          "numeroComprobante": {
            "type": "string",
            "nullable": true
          },
          "fechaContable": {
            "type": "string",
            "example": "2026-04-22"
          },
          "glosa": {
            "type": "string"
          },
          "glosaLinea": {
            "type": "string",
            "nullable": true
          },
          "estado": {
            "type": "string"
          },
          "anulado": {
            "type": "boolean"
          },
          "orden": {
            "type": "number"
          },
          "debeBob": {
            "type": "string",
            "example": "1000.00"
          },
          "haberBob": {
            "type": "string",
            "example": "0.00"
          },
          "saldoCorrienteBob": {
            "type": "string",
            "example": "1000.00"
          }
        },
        "required": [
          "comprobanteId",
          "numeroComprobante",
          "fechaContable",
          "glosa",
          "glosaLinea",
          "estado",
          "anulado",
          "orden",
          "debeBob",
          "haberBob",
          "saldoCorrienteBob"
        ]
      },
      "CuentaMayorDto": {
        "type": "object",
        "properties": {
          "cuentaId": {
            "type": "string"
          },
          "codigoInterno": {
            "type": "string"
          },
          "nombreCuenta": {
            "type": "string"
          },
          "naturaleza": {
            "type": "string"
          },
          "saldoInicialBob": {
            "type": "string",
            "example": "0.00"
          },
          "saldoFinalBob": {
            "type": "string",
            "example": "1000.00"
          },
          "totalDebeBob": {
            "type": "string",
            "example": "1000.00"
          },
          "totalHaberBob": {
            "type": "string",
            "example": "0.00"
          },
          "movimientos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MovimientoMayorDto"
            }
          }
        },
        "required": [
          "cuentaId",
          "codigoInterno",
          "nombreCuenta",
          "naturaleza",
          "saldoInicialBob",
          "saldoFinalBob",
          "totalDebeBob",
          "totalHaberBob",
          "movimientos"
        ]
      },
      "LibroMayorResponseDto": {
        "type": "object",
        "properties": {
          "rango": {
            "$ref": "#/components/schemas/RangoFechasDto"
          },
          "cuentas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CuentaMayorDto"
            }
          },
          "totalDebeBob": {
            "type": "string",
            "example": "1000.00"
          },
          "totalHaberBob": {
            "type": "string",
            "example": "1000.00"
          }
        },
        "required": [
          "rango",
          "cuentas",
          "totalDebeBob",
          "totalHaberBob"
        ]
      },
...
    }
  }
}